pyproject.toml
[tool.poetry]
name = "currantdjango"
version = "0.1.0"
description = ""
authors = ["Toan Phung <toan.phung@proteo.info>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
django = "^4.2.4"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.2.2"
djangorestframework = "^3.14.0"
alphastats = "^0.6.3"
gunicorn = "^21.2.0"
psycopg2 = "^2.9.7"
drf-url-filters = "^0.5.1"
drf-flex-fields = "^1.0.2"
drf-spectacular = "^0.26.4"
django-rest-auth = "^0.9.5"
dj-rest-auth = "^4.0.1"
django-request = "^1.6.3"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
boto3 = "^1.28.25"
django-storages = "^1.13.2"
channels-redis = {extras = ["cryptography"], version = "^4.1.0"}
django-dbbackup = "^4.0.2"
django-cors-headers = "^4.2.0"
django-rq = "^2.8.1"
django-redis = "^5.3.0"
rpy2 = "^3.5.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
